Added transparent backgrounds, fixed bugs.
The Blob bitmaps now use transparency. Probably need a way to toggle it for performance.
Changed fps for animation to be a Number so that you can do things like 0.5fps.
Some extra error checking for playAnimation.
A stub rotateTo function. Possibly useless.
Added rotation to convertFrameToBitmap, and fixed a bug or two in the counting of frames.